Add test_cp

+++ src/utilities/undo.py
@@ -41,4 +41,10 @@ def undo(flags: set, paths: list) -> None:
             os.remove(dest_path)
     else:
         for file_path, dest_path in paths:
-            shutil.move(dest_path, file_path)
+            if os.path.exists(dest_path):
+                print(f'{dest_path} пропущен: уже существует')
+                src.config.logger.main_logger.error(
+                    f'{dest_path} пропущен: уже существует'
+                )
+            else:
+                shutil.move(dest_path, file_path)
